+++ beep/models/__init__.py
@@ -9,6 +9,7 @@ from .base import ServerConfig, LevelOfTheory
 from .sampling import SamplingConfig
 from .sampling_periodic import SamplingPeriodicConfig
 from .be_comp_periodic import BeCompPeriodicConfig
+from .be_assemble_periodic import BeAssemblePeriodicConfig
 from .be_hess import BeHessConfig
 from .extract import ExtractConfig
 from .pre_exp import PreExpConfig